#118 Upravení designu okna pro správu kolekcí.
